Add pre-commit for code style checks
This adds the pre-commit framework for checking quality of committed
code and other files.

Additionally it fixes the parts of the code which didn't match the
configured rules. These changes are purely cosmetic and don't change
any logic in the code.

This commit also includes configuration for Github Actions to run the
checks for every pull request automatically.