Update supported Python versions
This updates the list of supported Python versions from 3.4, 3.5 and 3.6
to just 3.7, as that's the version currently used to run the lobby bots
for 0ad and is the only version of Python the bots are currently
properly tested with. Python 3.4 and 3.5 are also already end-of-life,
while Python 3.6 will reach its end-of-life in six months, so it makes
little sense to continue to support these versions.

While already supporting newer Python options would be nice, that'll
require some significant changes as development for SleekXMPP stopped
more than two years ago. There is a fork named Slixmmp as continuation,
but that fork requires at least Python 3.7. By updating the supported
Python versions, this commit lays the groundwork for replacing SleekXMPP
with Slixmpp and the support for newer Python versions.