Replace SleekXMPP with Slixmpp
SleekXMPP has been declared end-of-life for years and hasn't seen any
activity since. It also support only Python up to version 3.7. There is
a friendly fork called Slixmpp however, which sees continuous
development and releases and supports newer Python versions. To be able
to support newer Python versions with the lobby bots and to avoid using
abandoned libraries as dependencies, this replaces SleekXMPP with
Slixmpp.