Replace usage of ILIKE in database statements
When checking for player items in the leaderboard database, EcheLOn did
use ILIKE for some queries. This caused unexpected behavior which might
have lead to the wrong player item being returned from the database,
when a player JID contained an underscore, as underscores are handled as
single wildcard characters in LIKE-clauses.

This might have very well been the cause for the bug described in
https://trac.wildfiregames.com/ticket/6558.

This commit fixes that by replacing the use of ILIKE for
case-insensitive filtering of player JIDs with explicit filtering based
on the lower case player JID.