Remove use of deprecated SQLAlchemy features
This replaces the use of the [deprecated `backref` parameter][1] for
SQLAlchemy relationships and replaces it with explicit `back_populates`
definitions. It also replaces the many-to-many mapping using secondary
relationships with association proxies, as the former one was ambiguous
and caused SQLAlchemy to throw [conflicting column warnings][2].

With these changes EcheLOn should be compatible with the upcoming
SQLAlchemy 2.0.

[1]: https://docs.sqlalchemy.org/en/14/orm/backref.html
[2]: https://docs.sqlalchemy.org/en/14/errors.html#relationship-x-will-copy-column-q-to-column-p-which-conflicts-with-relationship-s-y