Add workaround for Slixmpp reconnect bug
This adds a workaround bug where subsequent calls to
ClientXMPP.connect() wouldn't use the same parameters as used by the
original call as intended by Slixmpp. A [patch with a fix for the
underlying bug][1] is available for Slixmpp, but until that's included
we need the workaround introduced in this commit.

[1]: https://lab.louiz.org/poezio/slixmpp/-/merge_requests/226