Don't show stale games anymore
Up until now we had the problem that stale games would be shown when
XpartaMuPP didn't receive the "got_offline" event when a player hosting
a game left the MUC. While it's unknown what causes XpartaMuPP not
receive such events from time to time, it's something which happened
regularly.

This commit fixes that by only returning games to players where
the hosting player is present in the MUC as well. As we don't know
if the hosting player stopped hosting a game or just had a short
disconnect from ejabberd, this just filters the list of returned games
and doesn't delete games in there. This should handle hosting players
temporarily dropping out of the MUC quite well and does only show their
game again once they rejoin and don't start hosting another game.