Add support for PubSub
This adds support for publishing game lists, rating lists and the
leader board via PubSub. With matching support in pyrogenesis this will
improve the performance and efficiency of the lobby, as the bots then
don't have to manually send updates to each connected client, but the
clients subscribe to the PubSub nodes and the XMPP server takes care of
deliverying updates instead.

The previous behavior of sending stanzas with updates to every connected
client is still available, but can be disabled with the
`--disable-legacy-lists` command line flag. The performance benefits
of PubSub do only materialize when the legacy behavior is disabled.