Always use TLS-encryption for connections
Up to now the bots supported unencryped connections to the XMPP server
when providing the `--disable-tls` command line option. This had two
use cases:

1. Testing with a local XMPP server where no TLS certificate signed by
   a trusted CA, but instead in most cases just a self-signed
   certificate, was available.
2. Connecting the bots through `localhost` as host name to the XMPP
   server, to ensure that traffic doesn't do a roundtrip over the
   public network, which did result in a mismatch of the host name of
   the TLS certificate and the actual host name (`localhost`).

As we want to transition to a setup with an XMPP server with unencrypted
connections being disabled for security and privacy reasons at some
point in future, we need to support these two use cases in a different
way. That's what this commit implements. Instead of offering the option
to disable the use of TLS for connections to the XMPP server, it adds an
option (`--no-verify`) to not verify the certificate presented by the
XMPP server. This allows connecting to an XMPP server which only offers
a self-signed certificate or whose host name doesn't match the one from
the certificate.

This change also has the benefit that communication between the bots
and the XMPP server is now always encrypted and a passive observer won't
be able to listen to it anymore.