Replace setup.py with pyproject.toml
This switches from specifying project metadata in setup.py to specifying
it in pyproject.toml, which is the recommended way to do so nowadays.
Doing so fixes a problem related to updating existing installations of
the bots. This wouldn't work previously, as the version number in
setup.py didn't change and pip therefore considered the code always the
same, no matter if there were actually code changes. This commit fixes
that by utilizing setuptools_scm to generate a version number
dynamically based on git tags and commits.