Add a new moderation bot
This adds a moderation bot as replacement of the existing bots in
https://gitlab.com/0ad/lobby-moderation

The new bot is a re-implementation from scratch focusing on complexity
reduction and code quality. This results in ~50% less code than before,
while keping the existing feature set, except for automated muting for
profanity, which wasn't used in the past months anyway.