Refactor logging for more granularity
This changes the use of the logger used for logging by the bots from the
root logger to a module specific logger, which provides more flexibility
in terms of configuration.

It also changes the command line options of the bots to configure
the verbosity of logging. This allows introducing an additional
verbosity level which uses debug logging for the bots themselves, but
doesn't do the same for third-party libraries used by them. This helps
reducing often undesired high-volume debug logs by libraries like
Slixmpp, while still generating debug logs for our custom code.