Remove Docker setup
Dockerfiles and instructions how to use Docker to locally run ejabberd
and the bots got added in the past to ease testing. After we did this,
we however added Ansible code to provision a real host as a Pyrogenesis
multiplayer lobby (https://github.com/0ad/lobby-infrastructure). This
code can also be used to deploy a multiplayer lobby to a virtual
machine utilizing Vagrant. The Ansible + Vagrant approach is much
easier to use than the Docker setup to get a test environment running
locally and resembles the real lobby setup much closer.

To avoid having to maintain two ways of how to run a local lobby for
testing, this removes the setup for Docker and points users to the
virtual machine approach.