Fix profile requests for invalid player names
Previously if a player used Pyrogenesis to look up the profile of
another player and typed in a player name, which doesn't resemble a
valid local part of a JID, EcheLOn would throw an exception. This commit
fixes that and properly returns the expected response when the queried
player doesn't exist.