Replace LimitedSizeDict with cachetools
When receiving game reports from Pyrogenesis, it regularly happens that
we never receive a report from both players of a rated game. This
happens in the following situations:

1. One of the player rage-quits by force-closing Pyrogenesis.
2. One player looses lobby connectivity during the game.
3. The rated game has only one human player and the other player is a
   bot (while Pyrogenesis shouldn't send reports for such games, here we
   are right now).

As we cache game reports in memory until we received and compared them
from both players of a rated game, this means that the `LimitedSizeDict`
used for that would reach its maximum size of 4096 items pretty fast.
While that doesn't impact functionality, it's using up way more memory
than necessary.

Additionally EcheLOn contained a bug which wouldn't even remove games
from the `LimitedSizeDict`, where we did receive all reports for.

While we can (and actually do in this commit) fix the bug in EcheLOn, we
have no control over the situations where we don't receive reports from
both players. However, what we can do is to remove reports for games
where we're pretty certain, that we'll never receive the missing report
anymore.

To do so, this commit replaces the custom `LimitedSizeDict` class with a
dependency on `cachetools`. For tracking reports from rated games a
size-limited `TTLCache` is now used which evicts reports belonging to
games after 1 hour, as we don't expect to receive reports later than a
few seconds after the end of a rated game.

This should result in a significant reduction of the memory required by
EcheLOn.