Use ruff for linting and formatting
This replaces linting with `prospector` and `bandit` with `ruff`, which
offers way more checks, is more actively developed and is _much_ faster.
This commit also adds a pre-commit hook to check for properly formatted
code using ruff as well.

While this commit touches most of the code, these are all non-functional
changes and only fix linting problems and properly format the code.