Add automatic checking for profanity
This adds functionality to automatically check for profanity in text
messages written in any of the XMPP MUC rooms monitored by the
moderation bot.

The terms being considered profanity can be configured using the
database and are language specific. They have to be stored in
their lemmatized form. If a supported language gets detected with an
accuracy of 100% only terms for that language will be checked, otherwise
English terms will be checked as well. Supported languages for now are
English, French, German, Polish, Portuguese, Russian, Spanish and
Turkish.

For the first two times in a sliding window of three months a user uses
profanity they'll receive a warning. Starting from the third time,
the user will get muted. At first users will be muted for five minutes,
with an exponentially increasing duration up to one week for each
continued use of profanity afterwards.

To enable this functionality the `--enable-profanity-monitoring`
command line option has to be provided.