Limit the length of game names to 256 characters
This limits the maximum length of game names to 256 characters. Longer
game names will be simply cut to the first 256 characters. While
Pyrogenesis already contains a check to limit the game name to 256
characters, this is to counter modified clients trying to circumvent
this limitation.

Fixes https://gitea.wildfiregames.com/0ad/0ad/issues/3651